Support out-of-workspace builds
Before this commit, watching of .su files and compile_commands.json
files outside of the opened workspace did not work. This was because
vscode.workspace.createFileSystemWatcher is not working for
out-of-workspace files when used with a string argument. Instead, the
argument should be a vscode.RelativePattern.

To also detect creation and deletion events, we need to provide an
appropriate base path to the RelativePattern constructor. It is natural
to use the directory of the compile commands entry. Because of this,
this commit is changing the signature of createFileSystemWatcher
to accept a baseDirectory and a pattern.