Minimize number of decoration setting
This commit significanty reduces the amount of times at which the
decorations for the visible editors are updated. This is based on two
core mechanism:

- When reacting to the onDidChangeVisibleTextEditors event, it is now
checked which of the visible text editors are actually new, i.e. have
not been visible before. Only for these we update the decorations.
- When updating the entries from an su file in the database, it is
checked which files (in particular headers) are affected by this change.
The decorations currently visible text editors are only updated if the
editors show one of the affected files.