Add workaround for GCC<10
As it turned out, gcc before version 10 does not put the complete
filename of a function into the .su file but only the basename (without
any directory information). This is not compatible to the current way of
searching for relevant results and also results in an error when calling
fs.realPath on it.

This commit resolves this issue by two main mechanisms:

- Replace pure basename .cpp filename entries with the known source
filename for the .su file.
- When querying for entries for a given file, also query for the
basename of the file. As these entries might also stem from other files
with identical basenames, try to validate each entry by checking if the
function name occurs in the respective document line.