Remove real path lookup for source file fixing
Using fs.realPathSync for source files bares the risk of asking for the
real path of a not yet existing path in cases where the source file is
generated at build time. Therefore, this commit temporarly disables
this lookup. Eventually, a multi-step watcher watching the link as well
as the potential link target will be required.